Bugfix for applet object tag.
Previous versions of processing failed to include all the necessary
jars in for the object 'archive' param element, so the object tag
needed to be tweaked (with some other additions).

With processing 1.5 though, all jars are included correctly.

Further, the 1.5 object tag fixes a 'Class not found' error when
accessing over the web, as opposed to running the applet locally.

We can now use the default processing 1.5 object tag, with the simple
addition of a 'codebase' param element pointing to the 'applet/' dir.