Latest. Replace applet with teaser image.
Processing 1.5 still has issues with OpenGL applets,
due in part to changes in JRE 6 update 21.

It looks like a solution is emerging with processing
rev 0198+, to be included in release 2.0.
(See processing issue 429.)

Since the desktop application still works fine
on all platforms, this replaces the demo applet
with a teaser image to those files for now.